split the offer test into different file
