+++ src/features/offer/pages/tests/OfferDescription.test.tsx
@@ -5,9 +5,9 @@ import React from 'react'
 import waitForExpect from 'wait-for-expect'
 
 import { OfferResponse } from 'api/gen'
+import { RootStack } from 'features/navigation/RootNavigator'
 import { offerResponseSnap } from 'features/offer/api/snaps/offerResponseSnap'
 import { dehumanizeId } from 'features/offer/services/dehumanizeId'
-import { RootStack } from 'features/navigation/RootNavigator'
 import { env } from 'libs/environment'
 import { ParsedDescription } from 'libs/parsers/highlightLinks'
 import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
